+++ Assets/Scripts/Model/Skill/WhiteSkill.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 
 public class WhiteSkill : Skill
 {
-    public WhiteSkill(int id, string name, SkillType type, int amount, float consume,
-        float damage, float def, string des, bool isEquip)
+    public WhiteSkill(int id, string name, SkillType type, int amount, int consume,
+        int damage, int def, string des, bool isEquip)
         : base(id, name, type, amount, consume, damage, def, des, isEquip)
     {
