+++ Assets/Scripts/Model/Skill/YellowSkill.cs
@@ -14,8 +14,8 @@ public class YellowSkill : Skill
     public float AddSpe { get; set; }
     public float AddLck { get; set; }
 
-    public YellowSkill(int id, string name, SkillType skillType, int amount, float consume,
-        float damage, float defense, string des, bool isEquip, string animId, float restore, float addHp, float addMp, float addAtk,
+    public YellowSkill(int id, string name, SkillType skillType, int amount, int consume,
+        int damage, int defense, string des, bool isEquip, string animId, float restore, float addHp, float addMp, float addAtk,
         float addDef, float reduceCon, float addSpe, float addLck)
         : base(id, name, skillType, amount, consume, damage, defense, des,isEquip)
     {
